output consistency: move creation of SqlExecutors
